+++ terminal.html
@@ -589,6 +589,6 @@
   <script src="node_modules/xterm-addon-search/lib/xterm-addon-search.js"></script>
   <script src="node_modules/xterm-addon-web-links/lib/xterm-addon-web-links.js"></script>
   <script src="node_modules/xterm-addon-serialize/lib/xterm-addon-serialize.js"></script>
-  <script src="src/js/terminal.js"></script>
+  <script src="src/js/terminal.js" type="module"></script>
 </body>
 </html>
